Store birthtimes as SQL NULL instead of 0 on systems that lack these
The value of this is irrelevant on those systems, but if a repo with a
ts.db is later moved to a system that does support birthtimes, those
will no longer be applied as the start of the epoch.

This also skips a redundant call to utimes on those systems.

dbs which never stored birthtimes may be retroactively updated with
`sqlite3 .git/ts.db "UPDATE timestamps SET birthtime = NULL WHERE birthtime = 0"`